Cleaning a mess, merging classes and rewriting bank account to store card informations in database
