+++ Simple Banking System/Simple Banking System/task/banking/banking.py
@@ -1,15 +1,21 @@
-from Bank import *
 import sqlite3
+from sqlite3 import Error
+from Bank import Bank
 
 
-# Only for passing test, all logic is in Database class
-db_connection = sqlite3.connect('card.s3db')
-cursor = db_connection.cursor()
-cursor.execute(""" CREATE TABLE IF NOT EXISTS card(
+try:
+    db_conn = sqlite3.connect('card.s3db')
+    cursor = db_conn.cursor()
+    cursor.execute(""" CREATE TABLE IF NOT EXISTS card(
                             id INTEGER primary key,
                             number TEXT NOT NULL,
                             pin TEXT NOT NULL,
                             balance INTEGER DEFAULT 0
                         );""")
-bank = Bank()
-bank.menu()
+    db_conn.commit()
+    bank = Bank(db_conn)
+    bank.menu()
+except Error as e:
+    print(e)
+finally:
+    db_conn.close()
